+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [5.2.1] - 2026-08-10
 ### Fixed
 - Call OnAnyAvailabilityChanged when InteractableHandler is destroyed
 
@@ -91,7 +92,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/1mbitshorde/InteractableSystem/compare/5.2.0...main
+[Unreleased]: https://github.com/1mbitshorde/InteractableSystem/compare/5.2.1...main
+[5.2.1]: https://github.com/1mbitshorde/InteractableSystem/tree/5.2.1/
 [5.2.0]: https://github.com/1mbitshorde/InteractableSystem/tree/5.2.0/
 [5.1.0]: https://github.com/1mbitshorde/InteractableSystem/tree/5.1.0/
 [5.0.0]: https://github.com/1mbitshorde/InteractableSystem/tree/5.0.0/